Groups all code relating to ingesting data into a KGCN into one module: 'ingest'
