+++ examples/kgcn/animal_trade/main.py
@@ -29,8 +29,8 @@ import kglib.kgcn.management.persistence as prs
 import kglib.kgcn.management.thing as thing_mgmt
 import kglib.kgcn.learn.classify as classify
 import kglib.kgcn.core.model as model
-import kglib.kgcn.neighbourhood.data.sampling.random_sampling as random_sampling
-import kglib.kgcn.preprocess.persistence as persistence
+import kglib.kgcn.core.ingest.neighbourhood.data.sampling.random_sampling as random_sampling
+import kglib.kgcn.core.ingest.preprocess.persistence as persistence
 
 flags = tf.app.flags
 FLAGS = flags.FLAGS
