+++ kglib/kgcn/core/ingest/encode/boolean.py
