+++ kglib/kgcn/core/ingest/encode/schema.py
