+++ kglib/kgcn/core/ingest/neighbourhood/data/sampling/ordered_test.py
@@ -1,3 +1,4 @@
+
 #
 #  Licensed to the Apache Software Foundation (ASF) under one
 #  or more contributor license agreements.  See the NOTICE file
@@ -19,7 +20,7 @@
 
 import unittest
 
-import kglib.kgcn.neighbourhood.data.sampling.random_sampling as rnd
+import kglib.kgcn.core.ingest.neighbourhood.data.sampling.ordered as ordered
 
 
 def _gen(l):
@@ -27,7 +28,7 @@ def _gen(l):
         yield i
 
 
-class TestRandomSample(unittest.TestCase):
+class TestOrderedSample(unittest.TestCase):
     def setUp(self):
         self._p = list(range(5))
 
@@ -35,9 +36,8 @@ class TestRandomSample(unittest.TestCase):
 
         population = _gen(self._p)
         n = 3
-        rnd_samples = rnd.random_sample(population, n)
-        self.assertEqual(min(n, len(self._p)), len(rnd_samples))
-        self.assertTrue(set(rnd_samples).issubset(set(self._p)))
+        first_n = ordered.ordered_sample(population, n)
+        self.assertListEqual(list(first_n), self._p[:n])
 
     def test_when_sample_size_more_than_population(self):
         ns = [7, 10, 11, 15, 24]
@@ -46,22 +46,26 @@ class TestRandomSample(unittest.TestCase):
             with self.subTest(f'population size = {len(self._p)}, n = {n}'):
                 population = _gen(self._p)
 
-                rnd_samples = rnd.random_sample(population, n)
-                self.assertEqual(min(n, len(self._p)), len(rnd_samples))
-                self.assertTrue(set(rnd_samples).issubset(set(self._p)))
+                first_n = ordered.ordered_sample(population, n)
+
+                self.assertListEqual(list(first_n), list(_gen(self._p))[:n])
 
     def test_when_sample_size_is_zero(self):
 
         population = _gen(self._p)
         n = 0
-        rnd_samples = rnd.random_sample(population, n)
-        self.assertListEqual(list(rnd_samples), [])
+        first_n = ordered.ordered_sample(population, n)
+        self.assertListEqual(list(first_n), [])
 
     def test_when_population_size_is_zero(self):
 
         population = _gen([])
         n = 5
         with self.assertRaises(ValueError) as context:
-            first_n = rnd.random_sample(population, n)
+            first_n = ordered.ordered_sample(population, n)
 
         self.assertTrue('Population is empty' in str(context.exception))
+
+
+if __name__ == "__main__":
+    unittest.main()
