+++ kglib/kgcn/core/ingest/neighbourhood/schema/traversal.py
