+++ kglib/kgcn/core/ingest/neighbourhood/schema/traversal_test.py
@@ -21,9 +21,9 @@ import unittest
 
 import grakn
 
-import kglib.kgcn.encode.encode as encode
-import kglib.kgcn.neighbourhood.schema.executor as ex
-import kglib.kgcn.neighbourhood.schema.traversal as trv
+import kglib.kgcn.core.ingest.encode.encode as encode
+import kglib.kgcn.core.ingest.neighbourhood.schema.executor as ex
+import kglib.kgcn.core.ingest.neighbourhood.schema.traversal as trv
 
 
 class TestGetSchemaConceptTypes(unittest.TestCase):
