+++ kglib/kgcn/core/ingest/preprocess/persistence.py
