+++ kglib/kgcn/core/ingest/preprocess/raw_array_builder.py
@@ -24,7 +24,7 @@ import collections
 import grakn
 import numpy as np
 
-from kglib.kgcn.neighbourhood.data import executor as data_ex, traversal as trv
+from kglib.kgcn.core.ingest.neighbourhood.data import traversal as trv, executor as data_ex
 
 
 def build_default_arrays(neighbourhood_sizes, n_starting_concepts, array_data_types):
